[PM-35100] Move NumberFormatter to reusable extension and localize billing and discount formats
